cloud_function_get_members/requirements.txt
discord.py==2.0.1
pandas==1.3.1
protobuf==4.21.6
pyarrow==5.0.0
google-cloud-storage>=1.41.0